Fixes - Round #2 [add] - Schema Guard module - Protects the migration and avoid dead loops on automatic deploys [changes] - migration files: updated to detect and work with potential issues or errors

+++ database/migrations/2018_11_07_192923_create_transfers_table.php
@@ -12,46 +12,48 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create($this->table(), function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->morphs('from');
-            $table->morphs('to');
-            $table
-                ->enum('status', ['exchange', 'transfer', 'paid', 'refund', 'gift'])
-                ->default('transfer');
+        if (! Schema::hasTable($this->table())) {
+            Schema::create($this->table(), function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->morphs('from');
+                $table->morphs('to');
+                $table
+                    ->enum('status', ['exchange', 'transfer', 'paid', 'refund', 'gift'])
+                    ->default('transfer');
 
-            $table
-                ->enum('status_last', ['exchange', 'transfer', 'paid', 'refund', 'gift'])
-                ->nullable();
+                $table
+                    ->enum('status_last', ['exchange', 'transfer', 'paid', 'refund', 'gift'])
+                    ->nullable();
 
-            $table->unsignedBigInteger('deposit_id');
-            $table->unsignedBigInteger('withdraw_id');
+                $table->unsignedBigInteger('deposit_id');
+                $table->unsignedBigInteger('withdraw_id');
 
-            $table->decimal('discount', 64, 0)
-                ->default(0);
+                $table->decimal('discount', 64, 0)
+                    ->default(0);
 
-            $table->decimal('fee', 64, 0)
-                ->default(0);
+                $table->decimal('fee', 64, 0)
+                    ->default(0);
 
-            $table->uuid('uuid')
-                ->unique();
-            $table->timestamps();
+                $table->uuid('uuid')
+                    ->unique();
+                $table->timestamps();
 
-            $table->foreign('deposit_id')
-                ->references('id')
-                ->on($this->transactionTable())
-                ->onDelete('cascade');
+                $table->foreign('deposit_id')
+                    ->references('id')
+                    ->on($this->transactionTable())
+                    ->onDelete('cascade');
 
-            $table->foreign('withdraw_id')
-                ->references('id')
-                ->on($this->transactionTable())
-                ->onDelete('cascade');
-        });
+                $table->foreign('withdraw_id')
+                    ->references('id')
+                    ->on($this->transactionTable())
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::drop($this->table());
+        Schema::dropIfExists($this->table());
     }
 
     private function table(): string
